// backend/infrastructure/cdk.out/asset.1762a12e3ba26a42a68633e7d52a2d312fbafd071435324019f0bdcfab7f25f9/types/api.ts
export interface CreateAccountRequest {
  username: string;
  handle: string; // @なし
  email: string;
  password: string;
  phone_number: string;
  account_type?: AccountType;
}

export interface CreateAccountResponse {
  account_id: string;
  handle: string;
  email: string;
}